refacotr(main.py, dependencies.py): move get_db func from main.py to dependencies.py
- move database session creation logic and related configurations from main.py to dependencies.py
